util/DwarfReader, ElfImporter: use unsigned addresses
This change alters DwarfReader and import-trace's ElfImporter so that they use
unsigned int for static address and line numbers instead of signed int.

+++ src/core/util/DwarfReader.cc
@@ -234,7 +234,7 @@ bool DwarfReader::read_mapping(std::string fileName, std::list<addrToLine>& addr
 			}
 
 			if (lineNo&&isCode) {
-				struct addrToLine newLine = { (int) addr, (int) lineNo, normalize(lineSource) };
+				struct addrToLine newLine = { addr, lineNo, normalize(lineSource) };
 				addrToLineList.push_back(newLine);
 			}
